refactored view package
