+++ OpenTapoWearOs/app/src/main/java/dev/veeso/opentapowearos/MainActivity.kt
@@ -18,10 +18,10 @@ import dev.veeso.opentapowearos.net.DeviceScanner
 import dev.veeso.opentapowearos.net.NetworkUtils
 import dev.veeso.opentapowearos.tapo.api.tplinkcloud.TpLinkCloudClient
 import dev.veeso.opentapowearos.tapo.device.Device
-import dev.veeso.opentapowearos.view.Credentials
-import dev.veeso.opentapowearos.view.DeviceCache
-import dev.veeso.opentapowearos.view.DeviceData
-import dev.veeso.opentapowearos.view.DeviceListAdapter
+import dev.veeso.opentapowearos.view.intent_data.Credentials
+import dev.veeso.opentapowearos.view.intent_data.DeviceCache
+import dev.veeso.opentapowearos.view.intent_data.DeviceData
+import dev.veeso.opentapowearos.view.main_activity.DeviceListAdapter
 import kotlinx.coroutines.*
 import java.net.Inet4Address
