+++ OpenTapoWearOs/app/src/main/java/dev/veeso/opentapowearos/view/device_activity/Color.kt
@@ -1,4 +1,4 @@
-package dev.veeso.opentapowearos.view
+package dev.veeso.opentapowearos.view.device_activity
 
 enum class Color {
     COOL_WHITE,
